# apps/backend-nest/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Vben Backend API",
    "description": "基于 NestJS + Zod 构建的现代化后端 API 服务\n\n## 架构特点\n- 使用 Zod 进行数据验证和类型定义\n- 统一的三段式返回格式 (code, message, data)\n- 自动生成的 OpenAPI 文档\n- 类型安全的 API 接口",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 认证令牌，格式：Bearer <token>"
      }
    }
  },
  "paths": {
...
    "/api/system/dept": {
      "post": {
        "summary": "创建部门",
        "description": "创建新的部门",
        "tags": [
          "部门管理"
        ],
        "operationId": "DepartmentController_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "部门名称"
                  },
                  "pid": {
                    "type": "string",
                    "nullable": true,
                    "description": "父级部门ID，支持树形结构"
                  },
                  "status": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1,
                    "default": 1,
                    "description": "部门状态：0-禁用，1-启用"
                  },
                  "remark": {
                    "type": "string",
                    "nullable": true,
                    "description": "备注信息"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "响应状态码，200表示成功，其他表示错误"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息，描述操作结果"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "部门唯一标识"
                        },
                        "name": {
                          "type": "string",
                          "maxLength": 100,
                          "description": "部门名称"
                        },
                        "pid": {
                          "type": "string",
                          "nullable": true,
                          "description": "父级部门ID，支持树形结构"
                        },
                        "status": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1,
                          "description": "部门状态：0-禁用，1-启用"
                        },
                        "remark": {
                          "type": "string",
                          "nullable": true,
                          "description": "备注信息"
                        },
                        "createTime": {
                          "type": "string",
                          "format": "date",
                          "description": "创建时间"
                        },
                        "updateTime": {
                          "type": "string",
                          "format": "date",
                          "description": "更新时间"
                        },
                        "parentId": {
                          "type": "string",
                          "nullable": true,
                          "description": "父级部门ID"
                        },
                        "childrenIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "子级部门ID列表"
                        },
                        "users": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "用户唯一标识符，使用 CUID 作为主键"
                              },
                              "username": {
                                "type": "string",
                                "maxLength": 50,
                                "description": "用户名，必须唯一，用于登录认证"
                              },
                              "email": {
                                "type": "string",
                                "maxLength": 100,
                                "format": "email",
                                "description": "用户邮箱地址，必须唯一，可用于找回密码等功能"
                              },
                              "name": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 50,
                                "description": "用户真实姓名"
                              },
                              "avatar": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 500,
                                "description": "用户头像 URL，可选字段，用于显示用户头像"
                              },
                              "deptId": {
                                "type": "string",
                                "nullable": true,
                                "description": "所属部门ID"
                              },
                              "status": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1,
                                "description": "用户账户状态：0-禁用，1-启用"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户创建时间，自动设置为当前时间"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户信息最后更新时间，每次更新时自动更新"
                              }
                            },
                            "required": [
                              "id",
                              "username",
                              "email",
                              "status",
                              "createdAt",
                              "updatedAt"
                            ]
                          },
                          "description": "部门下的用户"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "status",
                        "createTime",
                        "updateTime"
                      ],
                      "description": "响应数据"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/system/dept/list": {
      "get": {
        "summary": "获取部门列表",
        "description": "获取部门列表，支持筛选条件",
        "tags": [
          "部门管理"
        ],
        "operationId": "DepartmentController_findAll",
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "default": 1,
              "description": "页码，从1开始"
            },
            "required": false,
            "description": "页码，从1开始",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "description": "每页数量，最大100条"
            },
            "required": false,
            "description": "每页数量，最大100条",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "按部门名称筛选"
            },
            "required": false,
            "description": "按部门名称筛选",
            "name": "name",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1,
              "description": "按部门状态筛选：0-禁用，1-启用"
            },
            "required": false,
            "description": "按部门状态筛选：0-禁用，1-启用",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "按父级部门ID筛选"
            },
            "required": false,
            "description": "按父级部门ID筛选",
            "name": "pid",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "响应状态码，200表示成功，其他表示错误"
                    },
                    "message": {
                      "type": "string",
                      "description": "响应消息，描述操作结果"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "部门唯一标识"
                        },
                        "name": {
                          "type": "string",
                          "maxLength": 100,
                          "description": "部门名称"
                        },
                        "pid": {
                          "type": "string",
                          "nullable": true,
                          "description": "父级部门ID，支持树形结构"
                        },
                        "status": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1,
                          "description": "部门状态：0-禁用，1-启用"
                        },
                        "remark": {
                          "type": "string",
                          "nullable": true,
                          "description": "备注信息"
                        },
                        "createTime": {
                          "type": "string",
                          "format": "date",
                          "description": "创建时间"
                        },
                        "updateTime": {
                          "type": "string",
                          "format": "date",
                          "description": "更新时间"
                        },
                        "parentId": {
                          "type": "string",
                          "nullable": true,
                          "description": "父级部门ID"
                        },
                        "childrenIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "子级部门ID列表"
                        },
                        "users": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "description": "用户唯一标识符，使用 CUID 作为主键"
                              },
                              "username": {
                                "type": "string",
                                "maxLength": 50,
                                "description": "用户名，必须唯一，用于登录认证"
                              },
                              "email": {
                                "type": "string",
                                "maxLength": 100,
                                "format": "email",
                                "description": "用户邮箱地址，必须唯一，可用于找回密码等功能"
                              },
                              "name": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 50,
                                "description": "用户真实姓名"
                              },
                              "avatar": {
                                "type": "string",
                                "nullable": true,
                                "maxLength": 500,
                                "description": "用户头像 URL，可选字段，用于显示用户头像"
                              },
                              "deptId": {
                                "type": "string",
                                "nullable": true,
                                "description": "所属部门ID"
                              },
                              "status": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 1,
                                "description": "用户账户状态：0-禁用，1-启用"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户创建时间，自动设置为当前时间"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date",
                                "description": "用户信息最后更新时间，每次更新时自动更新"
                              }
                            },
                            "required": [
                              "id",
                              "username",
                              "email",
                              "status",
                              "createdAt",
                              "updatedAt"
                            ]
                          },
                          "description": "部门下的用户"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "status",
                        "createTime",
                        "updateTime"
                      ],
                      "description": "响应数据"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "请求参数验证失败"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}